Remove vault usage from Go code
Instead of asking the vault for the key in the go code, we get the key
in the filesystem somehow. This makes testing easier as we are then not
dependent on an extra vault instance. It also makes the tool easier to
deploy in diverse environments.

This key should be used for testing purposes of the tool.

Co-authored-by: jonandernovella <novella.uu@gmail.com>